requirements.txt
flask
flask-mysql-connector
setuptools>=70.0.0